+++ db_manager.py
@@ -5,15 +5,14 @@ A simple command-line tool to manage the database for AI News Bot.
 Provides commands to list entries, delete a specific entry by ID, or cleanup tables.
 """
 
-import sqlite3
 import argparse
 import logging
+import sqlite3
 import sys
-import os
+
+from settings import DB_PATH
 
 # --- Global Constants ---
-BASE_DIR = os.path.dirname(os.path.abspath(__file__))
-DB_NAME = os.path.join(BASE_DIR, "news_articles.db")
 TABLE_ARTICLES = "articles"
 TABLE_POSTS = "posts"
 
@@ -26,72 +25,57 @@ def list_entries(table):
     """
     List all entries from the specified table.
     """
-    conn = sqlite3.connect(DB_NAME)
-    cursor = conn.cursor()
-
     try:
-        cursor.execute(f"SELECT * FROM {table}")
-        rows = cursor.fetchall()
-        if not rows:
-            logger.info("No entries found in table '%s'.", table)
-            return
-
-        # Print column headers based on table name
-        if table == TABLE_ARTICLES:
-            headers = ["ID", "Title", "Summary", "Link", "Published At", "MD5Sum"]
-        elif table == TABLE_POSTS:
-            headers = ["ID", "MD5Sum"]
-        else:
-            headers = []
-        
-        # Print header row
-        print(" | ".join(headers))
-        print("-" * 80)
-        for row in rows:
-            # Convert all items to string for uniform display
-            row_str = " | ".join(str(item) for item in row)
-            print(row_str)
+        with sqlite3.connect(DB_PATH) as conn:
+            cursor = conn.cursor()
+            cursor.execute(f"SELECT * FROM {table}")
+            rows = cursor.fetchall()
     except sqlite3.Error as e:
         logger.error("Error listing entries from %s: %s", table, e)
-    finally:
-        conn.close()
+        return
+
+    if not rows:
+        logger.info("No entries found in table '%s'.", table)
+        return
+
+    if table == TABLE_ARTICLES:
+        headers = ["ID", "Title", "Summary", "Link", "Published At", "MD5Sum"]
+    elif table == TABLE_POSTS:
+        headers = ["ID", "MD5Sum"]
+    else:
+        headers = []
+
+    print(" | ".join(headers))
+    print("-" * 80)
+    for row in rows:
+        row_str = " | ".join(str(item) for item in row)
+        print(row_str)
 
 
 def delete_entry(table, entry_id):
-    """
-    Delete a specific entry by ID from the specified table.
-    """
-    conn = sqlite3.connect(DB_NAME)
-    cursor = conn.cursor()
-
+    """Delete a specific entry by ID from the specified table."""
     try:
-        cursor.execute(f"DELETE FROM {table} WHERE id = ?", (entry_id,))
-        if cursor.rowcount == 0:
-            logger.info("No entry with ID %d found in table '%s'.", entry_id, table)
-        else:
-            conn.commit()
-            logger.info("Deleted entry with ID %d from table '%s'.", entry_id, table)
+        with sqlite3.connect(DB_PATH) as conn:
+            cursor = conn.cursor()
+            cursor.execute(f"DELETE FROM {table} WHERE id = ?", (entry_id,))
+            if cursor.rowcount == 0:
+                logger.info("No entry with ID %d found in table '%s'.", entry_id, table)
+            else:
+                conn.commit()
+                logger.info("Deleted entry with ID %d from table '%s'.", entry_id, table)
     except sqlite3.Error as e:
         logger.error("Error deleting entry from %s: %s", table, e)
-    finally:
-        conn.close()
 
 
 def cleanup_table(table):
-    """
-    Delete all entries from the specified table.
-    """
-    conn = sqlite3.connect(DB_NAME)
-    cursor = conn.cursor()
-
+    """Delete all entries from the specified table."""
     try:
-        cursor.execute(f"DELETE FROM {table}")
-        conn.commit()
-        logger.info("Cleaned up all entries from table '%s'.", table)
+        with sqlite3.connect(DB_PATH) as conn:
+            conn.execute(f"DELETE FROM {table}")
+            conn.commit()
+            logger.info("Cleaned up all entries from table '%s'.", table)
     except sqlite3.Error as e:
         logger.error("Error cleaning up table %s: %s", table, e)
-    finally:
-        conn.close()
 
 
 def main():
