+++ main.py
@@ -4,12 +4,10 @@ main.py
 Responsible for generating social media posts for news articles and posting them to X/Twitter.
 """
 
-import configparser
 import hashlib
 import json
 import logging
 import re
-from dataclasses import dataclass
 from pathlib import Path
 from textwrap import dedent
 from typing import Optional, Tuple
@@ -17,48 +15,14 @@ from typing import Optional, Tuple
 import requests
 import sqlite3
 import tweepy
+from settings import DB_PATH, Config
 
 # Logging configuration
 logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(message)s")
 logger = logging.getLogger(__name__)
 
-# --- Paths ---
-BASE_DIR = Path(__file__).resolve().parent
-DB_PATH = BASE_DIR / "news_articles.db"
-CONFIG_PATH = BASE_DIR / "config.ini"
-
 # Regex to strip out <think>…</think>
 THINK_RE = re.compile(r"<think>.*?</think>", flags=re.DOTALL)
-
-
-@dataclass(frozen=True)
-class Config:
-    bearer_token: str
-    api_key: str
-    api_secret: str
-    access_token: str
-    access_token_secret: str
-    llm_api_url: str
-    model_name: str
-    debug_mode: bool
-
-    @classmethod
-    def load(cls, path: Path = CONFIG_PATH) -> "Config":
-        parser = configparser.ConfigParser(interpolation=None)
-        parser.read(path)
-        tw = parser["TwitterAPI"]
-        llm = parser["LLM"]
-        settings = parser["Settings"]
-        return cls(
-            bearer_token=tw["bearer_token"],
-            api_key=tw["api_key"],
-            api_secret=tw["api_secret"],
-            access_token=tw["access_token"],
-            access_token_secret=tw["access_token_secret"],
-            llm_api_url=llm["api_url"],
-            model_name=llm["model_name"],
-            debug_mode=settings.getboolean("debug_mode", fallback=False),
-        )
 
 
 def init_db(db_path: Path = DB_PATH) -> None:
